requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
numpy==1.26.2
scipy==1.11.4
pandas==2.1.3
pyarrow==14.0.1

# Options pricing
QuantLib-Python==1.32

# Machine learning (optional for future enhancements)
scikit-learn==1.3.2
statsmodels==0.14.0

# Task queue
celery==5.3.4
redis==5.0.1

# Caching
python-redis==0.1.4

# Authentication (for future)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

